% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylMix.R
\name{betaEst_2}
\alias{betaEst_2}
\title{The betaEst_2 function}
\usage{
betaEst_2(Y, w, weights)
}
\arguments{
\item{Y}{data vector.}

\item{w}{posterior weights.}

\item{weights}{Case weights.}
}
\value{
(a,b) parameters.
}
\description{
Internal. Estimates a beta distribution via Maximum Likelihood. Adapted from RPMM package.
}
\keyword{internal}

