% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylMix.R
\name{blc_2}
\alias{blc_2}
\title{The blc_2 function}
\usage{
blc_2(Y, w, maxiter = 25, tol = 1e-06, weights = NULL, verbose = TRUE)
}
\arguments{
\item{Y}{Data matrix (n x j) on which to perform clustering.}

\item{w}{Initial weight matrix (n x k) representing classification.}

\item{maxiter}{Maximum number of EM iterations.}

\item{tol}{Convergence tolerance.}

\item{weights}{Case weights.}

\item{verbose}{Verbose output.}
}
\value{
A list of parameters representing mixture model fit, including posterior weights and log-likelihood.
}
\description{
Internal. Fits a beta mixture model for any number of classes. Adapted from RPMM package.
}
\keyword{internal}

