% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylMix.R
\name{combineForEachOutput}
\alias{combineForEachOutput}
\title{The combineForEachOutput function}
\usage{
combineForEachOutput(out1, out2)
}
\arguments{
\item{out1}{result from one foreach loop.}

\item{out2}{result from another foreach loop.}
}
\value{
List with the combined results.
}
\description{
Internal. Function to combine results from the foreach loop.
}
\keyword{internal}

