% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{GEcancer}
\alias{GEcancer}
\title{Cancer Gene expression data of glioblastoma patients from the TCGA project}
\format{A numeric matrix with 14 rows (genes) and 251 columns (samples).}
\usage{
data(GEcancer)
}
\description{
Cancer Gene expression data of glioblastoma patients from the TCGA project. 
A set of 14 genes that have been shown in the literature to be involved in 
differential methylation in glioblastoma were selected as an example to try out MethylMix.
}
\references{
Cancer Genome Atlas Research Network. Comprehensive genomic characterization
defines human glioblastoma genes and core pathways. Nature. 2008 Oct 23;
455(7216):1061-8. doi: 10.1038/nature07385. Epub 2008 Sep 4. Erratum in: 
Nature. 2013 Feb 28;494(7438):506. PubMed PMID: 18772890; PubMed Central PMCID: PMC2671642.
}
\seealso{
TCGA: The Cancer Genome Atlas: \url{http://cancergenome.nih.gov/}
}
\keyword{datasets}

