\name{MAcancer}
\docType{data}
\alias{MAcancer}
\title{Cancer Gene expression data of glioblastoma patients from the TCGA 
project.}
\description{
     Cancer Gene expression data of glioblastoma patients from the TCGA project. 
     A set of 14 genes that have been shown in the literature to be involved in 
     differential methylation in glioblastoma were selected as an example to try 
     out MethylMix.
}
\usage{MAcancer}
\format{This describes the format of MAcancer.}
\source{This gives the source of MAcancer}
\references{
	Cancer Genome Atlas Research Network. Comprehensive genomic 
     characterization
     defines human glioblastoma genes and core pathways. Nature. 2008 Oct
     23;455(7216):1061-8. doi: 10.1038/nature07385. Epub 2008 Sep 4. Erratum in:
     Nature. 2013 Feb 28;494(7438):506. PubMed PMID: 18772890; PubMed Central PMCID:
     PMC2671642.
	
	Also see TCGA: The Cancer Genome Atlas: http://cancergenome.nih.gov/
}