\name{MAcancer_CDH1}
\docType{data}
\alias{MAcancer_CDH1}
\title{Cancer Gene expression data of CDH1 for breast cancer patients from the 
TCGA project.}
\description{
     Cancer Gene expression data for the gene CDH1 of breast cancer patients from 
     the TCGA project.
}
\usage{MAcancer_CDH1}
\format{This describes the format of MAcancer.}
\source{This gives the source of MAcancer}
\references{
	Cancer Genome Atlas Network. Comprehensive molecular portraits of human 
     breast tumours. 
     Nature. 2012 Oct 4;490(7418):61-70. doi: 10.1038/nature11412. Epub 2012 
     Sep 23. PubMed PMID: 23000897; PubMed Central PMCID: PMC3465532.	
	Also see TCGA: The Cancer Genome Atlas: http://cancergenome.nih.gov/
}