% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylMix.R
\name{MethylMix_RemoveFlipOver}
\alias{MethylMix_RemoveFlipOver}
\title{The MethylMix_RemoveFlipOver function}
\usage{
MethylMix_RemoveFlipOver(OrigOrder, MethylationState, classification,
  METdataVector, NrComponents, UseTrainedFlipOver = FALSE,
  FlipOverState = 0)
}
\arguments{
\item{OrigOrder}{order of sorted values in the methylation vector.}

\item{MethylationState}{methylation states for this gene.}

\item{classification}{vector with integers indicating to wich component each sample was classified into.}

\item{METdataVector}{vector with methylation values from the cancer samples.}

\item{NrComponents}{number of components in this gene.}

\item{UseTrainedFlipOver}{.}

\item{FlipOverState}{.}
}
\value{
Corrected vectors with methylation states and classification.
}
\description{
Internal. The estimated densities for each beta component can overlap, generating samples that look like being separated from their group.
This function re classifies such samples.
}
\keyword{internal}

