\name{METnormal}
\docType{data}
\alias{METnormal}
\title{DNA methylation data from normal tissue from glioblastoma patients.}
\description{
     Normal tissue DNA methylation data of glioblastoma patients. These normal 
     tissue samples were run on the same platform and are described in the      
     publication referenced below.
}
\usage{METnormal}
\format{This describes the format of METnormal.}
\source{This gives the source of METnormal}
\references{
     Noushmehr H, Weisenberger DJ, Diefes K, Phillips HS, Pujara K, Berman BP, 
     Pan F, Pelloski CE, Sulman EP, Bhat KP, Verhaak RG, Hoadley KA, Hayes DN, 
     Perou CM, Schmidt HK, Ding L, Wilson RK, Van Den Berg D, Shen H, 
     Bengtsson H, Neuvial P, Cope LM, Buckley J, Herman JG, Baylin SB, Laird 
     PW, Aldape K; Cancer Genome Atlas Research Network. Identification of a 
     CpG island methylator phenotype that defines a distinct subgroup of 
     glioma. Cancer Cell. 2010 May 18;17(5):510-22. doi: 
     10.1016/j.ccr.2010.03.017. Epub 2010 Apr 15. PubMed PMID: 20399149; PubMed
     Central PMCID: PMC2872684
}