\name{METnormal_CDH1}
\docType{data}
\alias{METnormal_CDH1}
\title{DNA methylation data for CDH1 from the breast cancer TCGA project.}
\description{
     Raw DNA methylation data for CDH1 from the 450k Infinium platform from breast
     cancer patients from the TCGA project.
}
\usage{METnormal_CDH1}
\format{This describes the format of METcancer.}
\source{This gives the source of METcancer}
\references{
	Cancer Genome Atlas Network. Comprehensive molecular portraits of human 
     breast tumours. 
     Nature. 2012 Oct 4;490(7418):61-70. doi: 10.1038/nature11412. Epub 2012 
     Sep 23. PubMed PMID: 23000897; PubMed Central PMCID: PMC3465532.	

	Also see TCGA: The Cancer Genome Atlas: http://cancergenome.nih.gov/
}
