% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethylMix.R
\name{predictOneGene}
\alias{predictOneGene}
\title{The predictOneGene function}
\usage{
predictOneGene(newVector, mixtureModel)
}
\arguments{
\item{newVector}{vector with new beta values}

\item{mixtureModel}{beta mixture model object for the gene being evaluated.}
}
\value{
A matrix with predictions (indices of mixture component), driver genes in rows, new samples in columns
}
\description{
Auxiliar function. Given a new vector of beta values, this function calculates a matrix with posterior prob of belonging 
to each mixture commponent (columns) for each new beta value (rows), 
and return the number of the mixture component with highest posterior probabilit
}
