% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{Preprocess_CancerSite_Methylation27k}
\alias{Preprocess_CancerSite_Methylation27k}
\title{The Preprocess_CancerSite_Methylation27k function}
\usage{
Preprocess_CancerSite_Methylation27k(CancerSite, METdirectory,
  MissingValueThreshold = 0.2)
}
\arguments{
\item{CancerSite}{character of length 1 with TCGA cancer code.}

\item{METdirectory}{character with directory where a folder for downloaded files will be created. Can be the object returned by the Download_DNAmethylation function.}

\item{MissingValueThreshold}{threshold for removing samples or genes with missing values.}
}
\value{
List with pre processed methylation data for cancer and normal samples.
}
\description{
Internal. Pre-processes DNA methylation data from TCGA from Illymina 27k arrays.
}
\keyword{internal}

