% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ProbeAnnotation}
\alias{ProbeAnnotation}
\title{ProbeAnnotation data set}
\description{
Data set with annotation from Illumina methylatin arrays mapping CpG sites to genes.
}
\keyword{datasets}

