% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{TCGA_BatchCorrection_MolecularData}
\alias{TCGA_BatchCorrection_MolecularData}
\title{The TCGA_BatchCorrection_MolecularData function}
\usage{
TCGA_BatchCorrection_MolecularData(GEN_Data, BatchData, MinInBatch)
}
\arguments{
\item{GEN_Data}{matrix with data to be corrected for batch effects.}

\item{BatchData}{Batch data.}

\item{MinInBatch}{minimum number of samples per batch.}
}
\value{
The corrected data matrix.
}
\description{
Internal. Wrapper to perform batch correction.
}
\keyword{internal}

