% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{TCGA_GENERIC_BatchCorrection}
\alias{TCGA_GENERIC_BatchCorrection}
\title{The TCGA_GENERIC_BatchCorrection function}
\usage{
TCGA_GENERIC_BatchCorrection(GEN_Data, BatchData)
}
\arguments{
\item{GEN_Data}{matrix with data to be corrected for batch effects.}

\item{BatchData}{Batch data.}
}
\value{
The corrected data matrix.
}
\description{
Internal. Performs batch correction.
}
\keyword{internal}

