% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{TCGA_GENERIC_CheckBatchEffect}
\alias{TCGA_GENERIC_CheckBatchEffect}
\title{The TCGA_GENERIC_CheckBatchEffect function}
\usage{
TCGA_GENERIC_CheckBatchEffect(GEN_Data, BatchData)
}
\arguments{
\item{GEN_Data}{matrix with data to be corrected for batch effects.}

\item{BatchData}{Batch data.}
}
\value{
list with results.
}
\description{
Internal. Checks if batch correction is needed.
}
\keyword{internal}

