% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{TCGA_GENERIC_CleanUpSampleNames}
\alias{TCGA_GENERIC_CleanUpSampleNames}
\title{The TCGA_GENERIC_CleanUpSampleNames function}
\usage{
TCGA_GENERIC_CleanUpSampleNames(GEN_Data, IDlength = 12)
}
\arguments{
\item{GEN_Data}{data matrix.}

\item{IDlength}{length of samples ID.}
}
\value{
data matrix with cleaned sample names.
}
\description{
Internal. Cleans the samples IDs into the 12 digit format and removes doubles.
}
\keyword{internal}

