% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{TCGA_GENERIC_GetSampleGroups}
\alias{TCGA_GENERIC_GetSampleGroups}
\title{The TCGA_GENERIC_GetSampleGroups function}
\usage{
TCGA_GENERIC_GetSampleGroups(SampleNames)
}
\arguments{
\item{SampleNames}{vector with sample names.}
}
\value{
a list.
}
\description{
Internal. Looks for the group of the samples (normal/cancer).
}
\keyword{internal}

