% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{TCGA_GENERIC_LoadIlluminaMethylationData}
\alias{TCGA_GENERIC_LoadIlluminaMethylationData}
\title{The TCGA_GENERIC_LoadIlluminaMethylationData function}
\usage{
TCGA_GENERIC_LoadIlluminaMethylationData(Filename)
}
\arguments{
\item{Filename}{name of the file with the data.}
}
\value{
methylation data.
}
\description{
Internal. Read in an illumina methylation file with the following format: header row with sample labels, 
2nd header row with 4 columns per sample: beta-value, geneSymbol, chromosome and GenomicCoordinate. 
The first column has the probe names.
}
\keyword{internal}

