% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{TCGA_GENERIC_MergeData}
\alias{TCGA_GENERIC_MergeData}
\title{The TCGA_GENERIC_MergeData function}
\usage{
TCGA_GENERIC_MergeData(NewIDListUnique, DataMatrix)
}
\arguments{
\item{NewIDListUnique}{unique rownames of data.}

\item{DataMatrix}{data matrix.}
}
\value{
data matrix.
}
\description{
Internal.
}
\keyword{internal}

