% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{TCGA_GENERIC_MET_ClusterProbes_Helper_ClusterGenes_with_hclust}
\alias{TCGA_GENERIC_MET_ClusterProbes_Helper_ClusterGenes_with_hclust}
\title{The TCGA_GENERIC_MET_ClusterProbes_Helper_ClusterGenes_with_hclust function}
\usage{
TCGA_GENERIC_MET_ClusterProbes_Helper_ClusterGenes_with_hclust(Gene,
  ProbeAnnotation, MET_Cancer, MET_Normal = NULL, CorThreshold = 0.4)
}
\arguments{
\item{Gene}{gene.}

\item{ProbeAnnotation}{data set matching probes to genes.}

\item{MET_Cancer}{data matrix for cancer samples.}

\item{MET_Normal}{data matrix for normal samples.}

\item{CorThreshold}{correlation threshold for cutting the clusters.}
}
\value{
List with the clustered data sets and the mapping between probes and genes.
}
\description{
Internal. Cluster probes into genes.
}
\keyword{internal}

