% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{TCGA_Load_MolecularData}
\alias{TCGA_Load_MolecularData}
\title{The TCGA_Load_MolecularData function}
\usage{
TCGA_Load_MolecularData(Filename, MissingValueThresholdGene = 0.3,
  MissingValueThresholdSample = 0.1)
}
\arguments{
\item{Filename}{name of the file with the data.}

\item{MissingValueThresholdGene}{threshold for missing values per gene. Genes with a percentage of NAs greater than this threshold are removed. Default is 0.3.}

\item{MissingValueThresholdSample}{threshold for missing values per sample. Samples with a percentage of NAs greater than this threshold are removed. Default is 0.1.}
}
\value{
gene expression data.
}
\description{
Internal. Reads in gene expressiondata. Deletes samples and genes with more NAs than the respective thresholds. Imputes other NAs values.
}
\keyword{internal}

