% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_Preprocess.R
\name{TCGA_Process_EstimateMissingValues}
\alias{TCGA_Process_EstimateMissingValues}
\title{The TCGA_Process_EstimateMissingValues function}
\usage{
TCGA_Process_EstimateMissingValues(MET_Data, MissingValueThreshold = 0.2)
}
\arguments{
\item{MET_Data}{data matrix.}

\item{MissingValueThreshold}{threshold for removing samples and genes with too many missing values.}
}
\value{
the data set with imputed values and possibly some genes or samples deleted.
}
\description{
Internal. Removes patients and genes with more missing values than the MissingValueThreshold, and imputes remaining missing values using Tibshirani's KNN method.
}
\keyword{internal}

